SpringBoot/src/main/java/com/example/springboot/service/impl/UserServiceImpl.java
package com.example.springboot.service.impl;

import com.example.springboot.domain.Role;
import com.example.springboot.domain.User;
import com.example.springboot.repos.RoleRepo;
import com.example.springboot.repos.UserRepo;
import com.example.springboot.service.UserService;
import lombok.NonNull;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.ArrayList;
import java.util.Collection;
import java.util.List;

@Service
@RequiredArgsConstructor
@Slf4j
@Transactional
public class UserServiceImpl implements UserService, UserDetailsService {

    private final UserRepo userRepo;
    private final RoleRepo roleRepo;
    private final PasswordEncoder passwordEncoder;

    @Override
    public User saveUser(@NonNull User user) {
        user.setPassword(
                passwordEncoder.encode(user.getPassword()));
        log.info("Save user: {}", user);
        return userRepo.save(user);
    }

    @Override
    public Role saveRole(@NonNull Role role) {
        log.info("Save role: {}", role);
        return roleRepo.save(role);
    }

    @Override
    public void addRoleToUser(String username, String roleName) {
        User user = userRepo.findByUsername(username);
        Role role = roleRepo.findByName(roleName);
        user.getRoles().add(role);
        userRepo.save(user);
        log.info("Added user role {} for user: {}", roleName, username);
    }

    @Override
    public User getUser(String username) {
        log.info("Find user by name: {}", username);
        return userRepo.findByUsername(username);
    }

    @Override
    public List<User> getUsers() {
        log.info("Get all users. NUm of entities {}.", userRepo.count());
        return userRepo.findAll();
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepo.findByUsername(username);
        if (user == null) {
            log.error("User with username {} not found!", username);
            throw new UsernameNotFoundException("User not found!");
        }
        log.info("User found in DB with username: {}", username);

        Collection<SimpleGrantedAuthority> authorities = new ArrayList<>();
        user.getRoles()
                .forEach(role -> authorities.add(new SimpleGrantedAuthority(role.getName())));
        return new org.springframework.security.core.userdetails.User(
                user.getUsername(), user.getPassword(), authorities);
    }
}
